refactor: moved MandatoryInfoBox
